Add half-session deduction for leave (ลา หัก 0.5 ครั้ง)
New orange "ลา½" button in check-in, alongside the existing full leave.

Model: Attendance gains deducted_units (1 / 0.5 / 0) derived from status via
a single DEDUCT_UNITS map, and Enrollment.used_sessions() now sums those
units instead of counting rows, so remaining_sessions is a Decimal.
Existing rows are backfilled from the old boolean in the migration.

Because remaining is now fractional, every int(remaining_sessions) had to
go: int() silently turns 1.5 into 1, which would have quietly understated
what a parent has left in twelve different places (dashboard AJAX, exports,
JSON payloads, the renewal list). Templates render counts through a new
|sessions filter so they read "1.5" and "10", never "1.0".

Renewal notices now trigger at 1.5 rather than 1 (RENEWAL_NOTICE_REMAINING_MAX),
since a course ending on a half-day leave is just as finished. Pricing is
untouched: a trailing 0.5 is not billed, and both the notice and the
attendance card now say the parent gets one catch-up clip for it instead.

The attendance card counts in units too -- it walks rows newest-first until
their units reach the used total. Half-day leave shows a ½ in the sequence
rather than a number, and the remaining dots end in a half dot.

Also swept the places a new status would otherwise have vanished from:
headcount summaries group it with ลา, the check-in tally no longer reports
it as "not yet checked", the ลา filter chip matches it, and anything that
turns attendance into money (finance estimate, school overview, recognised
revenue) counts it as the half session it actually is.

+++ core/exports.py
@@ -153,7 +153,7 @@ def build_full_workbook() -> tuple[BytesIO, str]:
     def _student_rows():
         for s in students:
             remaining = sum(
-                int(e.remaining_sessions or 0)
+                float(e.remaining_sessions or 0)
                 for e in s.enrollments.all()
                 if e.is_active
             )
@@ -186,7 +186,7 @@ def build_full_workbook() -> tuple[BytesIO, str]:
              getattr(e.tutoring_class, "name", "") or "",
              e.get_enrollment_type_display() if hasattr(e, "get_enrollment_type_display") else "",
              int(e.sessions_total or 0),
-             int(e.remaining_sessions or 0) if e.remaining_sessions is not None else 0,
+             float(e.remaining_sessions or 0) if e.remaining_sessions is not None else 0,
              _money(e.course_price), _money(e.discount_amount), _money(e.net_price),
              bool(e.is_active), _fmt_dt(e.created_at)]
             for e in enrollments
